WL7716: Move commercial files and distribution to CPYINT
The files related to the commercial distribution move to a new
internal repository (known as CPYINT). We also move all custom
Distutils commands for the creation of packages distributed by Oracle
to CPYINT.

If the Git submodule cpyint is available, setup.py has access the
custom Distutils commands to create GPL and commercial packages.

Overview of files which we rename, remove or move to the CPYINT
repository:

  AUTHORS -> (removed)
  ChangeLog -> CHANGES.txt
  README -> README.txt
  COPYING -> LICENSE.txt
  README_com.txt -> cpyint/data/commercial/README_COM.txt
  LICENSE_com.txt -> cpyint/data/commercial/LICENSE_COM.txt
  metasetupinfo.py -> setupinfo.py and cpyint/metasetupinfo.py
  docs/mysql* -> (removed)
  support/Debian -> cpyint/data/Debian
  support/RPM -> cpyint/data/RPM
  support/OSX -> cpyint/data/OSX
  support/MSWindows -> cpyint/data/MSWindows
  support/scripts/* -> cpyint/scripts/
  version.py -> python2/mysql/connector/version.py
                python3/mysql/connector/version.py