WL7462: Consolidate Python v2 and v3 code
We consolidate the Python v2 and v3 code into a new
location lib/ taking the Python v3 code as base.

Overview of files which we rename, remove or move:
  python3/mysql -> lib/mysql
  python23/django -> lib/mysql/connector/django
  python23/fabric -> lib/mysql/connector/fabric
  python3/examples -> examples/
  python2/ -> (removed)
  tests/py3/test_*.py -> tests/
  tests/py2 -> (removed)
  tests/py3 -> (removed)

The catch23 module is added to hold variables and functions to
make it easier to keep compatibility with Python v2 and
v3.

Some code is refactored to resolve some performance issues
introduced by making Python v3 code work better with v2.
For example the BaseSocket.recv() method is using memoryview
for versions of Python v2.7 and greater.

We add a new module custom_types which currently only hold
the HexLiteral type.

We remove the config.py file from the example/ folder. The
configuration is now included in each example script.

The linting issues are written to pylint_output.txt in the
root of the repository.