WL7292: Add cursors returning dict and namedtuple as rows
We implement new cursor classes returning rows as dictionaries
or named tuples and are available as both buffered and
non-buffered.

The new classes in mysql.connector.cursor are:
 * MySQLCursorDict
 * MySQLCursorBufferedDict
 * MySQLCursorNamedTuple
 * MySQLCursorBufferedNamedTuple

A raw version for the above classes is not available.

The cursor()-method of MySQLConnection has two new arguments
dictionary and named_tuple which can be set to True, for example,
to instantiate a buffered cursor returning rows as dict-objects:

  cur = cnx.cursor(dictionary=True, buffered=True)