WL7643: Add Python C Extension using Connector/C
We add a Python C Extension which uses MySQL Connector/C
and can be optionally compiled to improve, in some cases,
performances.

When the C Extension is compiled and installed, it will
be used instead of the pure Python connector. Although
care has been taken so the C Extension works the same
way as the pure Python solution, there could be
differences, for example, some error messages will be
different or even be more accurate.

To compile and install the C Extension, make sure
MySQL Connector/C or MySQL Client libraries is installed
and use the --with-mysql-capi to provide the absolute
path of mysql_config. On Windows, you need to provide
the folder where there Client libraries are installed.
For example:

 python setup.py install \
   --with-mysql-capi=/opt/mysql-connector-c/bin/mysql_config

The above command will install the _mysql_connector module
inside the Python site-packages folder.

By default, even though the C Extension is available, it is
not used unless the connection argument 'use_pure' is set
to False.
  cnx = mysql.connector.connect(.., use_pure=False)

The current C Extension does not support MySQL prepared
statements using the binary protocol.

Lots of unit tests have been manipulated so they are being
tested using both the pure connector and the C Extension.
To test both, it is needed to provided the
--with-mysql-capi option for unittests.py

  python unittests.py ... \
    --with-mysql-capi=/path/to/mysql_config