WL7937: Allow LOAD DATA LOCAL INFILE by default
We make the use of LOAD DATA LOCAL INFILE easier by enabling
the client flag 'LOCAL_FILES' by default when the client connects.
Previously the flag had to be set explicitly set using the
client_flags argument.

A new connection argument 'allow_local_infile', defaulting
to True, is added. (Note that the MySQL server still needs to be
configured to allow this feature.)

Unit tests are updated and a new test case was added to
test_connection.py.