WL7228: Allow Connector/Python to use MySQL option files
We add a new connection argument 'option_files' with which the
user can provide a list of MySQL option files. These files are
read in order given.

  cnx = mysql.connector.connect(option_files='/etc/mysql/cpy.cnf')

  mysql_option_files = [
    '/etc/mysql/cpy.cnf',
    './development.cnf',
  ]
  cnx = mysql.connector.connect(option_files=mysql_option_files)

Any connection argument given when instantiating a MySQLConnection
object takes precedence over configuration found in the option files.

By default, Connector/Python reads the groups 'client' and
'connector_python'. The user can override this by specifying an
additional connection argument which lists which groups to read.

The following example makes sure that only the 'connector_python'
group is read from the default locations:

  cnx = mysql.connector.connect(option_files=='/etc/mysql/cpy.cnf',
                                option_groups='connector_python')

The order in which groups are read will be defined by the order of
how they were given with the 'option_groups' argument. In the
following example, first the 'connectors' groups is read, and
then 'connector_python':

  groups = ['connectors', 'connector_python']
  cnx = mysql.connector.connect(
    option_files='/etc/mysql/cpy.cnf'
    option_groups=groups)

The value of an option is overwritten if the same option is found in
the next group given in the list.