BUG19584051: Fix comparision of type_code of columns for PEP-249
The type_code in cursor.description was not comparing equal
to any of the type objects defined in mysql.connector.dbapi.
We fix the issue by removing __cmp__ and implementing the
rich comparison operators for the PEP-249 class
DBAPITypeObject.

A unit test has been added for BUG#19584051.
(cherry picked from commit f985cfc)