BUG19500097: Fix string decoding with binary character set
When connection character set is set to 'binary', the string data
failed to decode since it is not a valid Python character set.
We fix this by returning string data without decoding when the
character set is set to 'binary'.

A unit test has been added for BUG#19500097.
(cherry picked from commit dcd5b63)