BUG19331658: Fix connection pooling with MySQL Fabric
Using connection pooling with MySQL Fabric was raising AttributeError.
We fix this by creating a pool of connections to a server as and when
first time it is required to be connected.

To use pooling with Fabric, 'pool_size' argument should be used. Using
'pool_name' with Fabric raises AttributeError. This is because
we can't have two pools with same names on different servers.

A unit test has been added for BUG#19331658.