BUG20022533: Fix failing Django inspectdb command with C/Python
Django inspectdb command was failing with "unread result found error".
The reason was newly added "consume_results" option was not behaving
as expected, raising this error. We fix this by adding "handle_unread_result"
method to connection module and using it whenever a check for unread
results are made, consuming all remaining results if "consume_results"
is set to True.

Unit tests are updated.