BUG19500097: Fix regression caused by C Extension
BUG#19500097 resurfaced with the merging of C Extension with the
develop branch. We fix the regression and also fix the same problem
of 'binary' character set in C Extension.

The patch also fixes 'foreach_cnx' decorator in the tests, which was
consuming all the Exceptions not allowing any test case failure to
show up.
We also fix set_charset_collation() method which was not setting
the character set of the current connection properly.

Unit tests are updated.