BUG20114310: Add support static linking C Extension
We add support for statically linking C libraries into
the C extension. A new command 'build_ext_static' was
added and works similar to 'build_ext' (which will build
dynamically).

The 'install' command accepts a new option --static.

This is most useful for packaging and/or when you do not
want to have dependencies on, for example, the MySQL C
libraries.