BUG19972427: Fix creating of redundant connections in Django
The Connector/Python Django backend was creating too many database
connections since as soon as an object of DatabaseWrapper was
instanciated, a connection to the MySQL server was set up. This is
required to enable some version specific MySQL features.

We fix this by adding the cached_property decorator to the
mysql_version()-method.

Unit tests are updated and also fixes BUG#19954882.