WL7956: Support MySQL Protocol connecting to MySQL Fabric
We add support for connecting to MySQL Fabric using the MySQL
protocol. Previously, it was only possible to use XMLRPC.

A new connection argument 'protocol' for the MySQL Fabric
configuration is made available. It accepts the values
'xmlrpc' or 'mysql'. Any other value would result in an
InterfaceError to be raised.

The default TCP port for the MySQL protocol of MySQL Fabric
is 32275 (XMLRPC is 32274).

The default protocol used is still XMLRPC.

Example usage:

  fabric_config = {
    'host': 'fabric.example.com',
    'protocol': 'mysql',
  }
  cnx = mysql.connector.connect(fabric=fabric_config, ..

Currently, when running Fabric unit tests for Connector/Python,
the tests will use the XMLRPC protocol. Use the unittests.py
command line option --with-fabric-protocol to change:

  $ python unittests.py --with-fabric-protocol=mysql ...
  ..INFO] Fabric will be tested using the 'MYSQL' protocol