BUG21490865: Fix compiling CExtension with relocated libmysqlclient
Install fails with --with-mysql-capi option when libmysqlclient
directory is renamed and the path does not contain '64' on 64 bit
systems. The reason being file command without '-L' option doesn't
derefer the symbolic links and doesn't print the architecture of the
original file.

We fix this by including '-L' option while using file command also we
now don't check the path of the file to contain '64'.

(cherry picked from commit 6475c80)