BUG21536507: Fix raising warnings as exceptions
We fix a few issues around raising_on_warnings is set to True. When
using CExtension, no errors were thrown as exceptions when the
executed statement produced an error. Also, when statement produces a
result, it was not possible to reuse the cursor.

A test case has been added for BUG#21536507
(cherry picked from commit 6a2b67d)