BUG21535573: Fix character decoding of identifiers using CExtension
We fix in CExtension the decoding of column names, table name and
database name in field information of result sets. When the decoding
fails, an error is now raised (instead of a segmentation fault).

A test case was added for BUG21535573 which is testing GBK, SJIS and
BIG5 characters in column names.
We also update a few other tests cases.
(cherry picked from commit 6e4649c)