BUG22545879 Fix usage of --ssl-cipher option
Previously, it was not possible to set the SSL ciphers option when
using SSL connections. We fix this by allowing a new connection
argument `ssl_cipher` and also using the MySQL option `--ssl-cipher`
when it is specified in option files.

Setting the SSL cipher can be important when facing the following
error message returned by OpenSSL:

 :SSLroutines:SSL3_CHECK_CERT_AND_ALGORITHM:dh key too small

Setting the cipher to 'AES256-SHA' for example, will help.

A test case for BUG22545879 was added. Multiple test cases were also
updated to a cipher.