BUG21449207 Fix fetching compressed data bigger than 16M
Connector/Python was stuck receiving from network when receiving
compressed data bigger than 16M.

We fix this by correctly stopping to receive data from network when
no more compressed data is send by the server.

A test case for bug 21449207 was added as well as adding more cases
while testing compressed packets. A test was removed in
`test_network.py`: it was not correctly done and code is checked with
other tests.