BUG23567257: Fix _id auto-generation
On inserting a document into a collection without specifying the `_id`
attribute, the following error is thrown:

Data too long for column '_id' at row 1

UUID generates a 36 character long string but `_id` can only take a
maximum of 32 characters. Generating the UUID as a 32-character
hexadecimal string fixes the issue.

The existing tests for `Collection.add` will pass with this fix.