Add View support (without DDL)
Given the Views behave similar to Tables on CRUD operations, they
are treated as if they were tables.

This patch insures that all operations available through a Table
object are also available if the underlying object is a View.

Table.is_view() method was added to verify if the underlying object
is a View.

Tests were added/changed for regression.