BUG23722871: Fix running unittests with mysqlx plugin
Unittests hang when running the entire test suite with MySQLx tests.

Each of the mysqlx tests in mysqlx_crud attempts to drop the main schema
`myconnpy` as a part of test tearDown method. This causes a database
lock since there are other sessions open with `myconnpy` as the default
database.

Previous sessions need to be closed before attempting to drop the schema.
But the Session.close API is included in release 2.2.1. As of now, we do
nothing in the tear down method.