BUG24946101: Fix inline foreign key SQL generation.
This patch fixes the inline foreign key reference SQL generation. A
non-existent variable name was used in place of the reference table
name which gave rise to this issue.

Tests cannot be verified because according to MySQL documentation:

MySQL parses but ignores "inline REFERENCES specifications" (as defined
in the SQL standard) where the references are defined as part of the
column specification. MySQL accepts REFERENCES clauses only when specified
as part of a separate FOREIGN KEY specification.