BUG24947078: Create or alter view raises an error using non updatable views
This patch fixes the error raised when attempting to create or alter non
updatable views.

By default the statement for creating or altering views included WITH CHECK
OPTION clause, this was removed to allow the usage of non updatable views.
The user can still use the with_check_option method to set this clause.

Tests were added for regression.