BUG21476495: Fix set_charset_collation() for an invalid charset provided
Incase of an integer value of the charset, CharacterSet.get_charset_info()
validates the charset and throws an error, but the value was assigned to
charset_id variable of the Connection class before being validated.

Assigning only after validating the charset id resolves this issue.

Unittests have been added for regression.