BUG21477493: Fix RE_SQL_INSERT_STMT to correctly match Insert Statement
The regex matches the literal "VALUES" in the insert string and assumes
that it is a batch insert, which may not be the case if this literal is
used as a value itself.

The regex is modified to match the complete Insert Statement and made sure
it conforms to the format `INSERT INTO [table_name] (col,...) VALUES(a,...)`.

Unittests have been added for regression.