WL#10081: DevAPI IPv6 Support
This feature adds IPv6 support to Connector/Python.
Currently, the user can connect to a MySQL server running the XPlugin using
a hostname that resolves to an IPv4 address.
With this feature, the hostname maybe resolved as IPv4 or IPv6 address.

This patch also updates the address list parsing from URI string to correctly
parse IPv6 addresses.