BUG25383644: Add connection back to pool on exception
Each time the connection to the server is lost during a query, the connection
is not added back to the connection pool. Eventually, there are no connections
left in the connection pool and the following error message is thrown:
mysql.connector.errors.PoolError: Failed getting connection; pool exhausted

To fix this error, we add the connection back to the pool even if the closing
of the connection was unsuccessful.

Tests added for regression.