BUG25558885: Set default connection timeout to pure connector/python
With any long running queries, the cext connector/python would lose
connection to the server with the message:
ERROR 2013 (LOST CONNECTION TO MYSQL SERVER)

But the pure connector/python does not throw any such error.

To make the behaviour consistent, we set the `connection_timeout` value
to None.

Test added for regression.