WL10658: Remove concept of NodeSession
The initial version of the X DevAPI was build around the concept of two
different session types, XSession and NodeSession.

This worklog removes the concept of NodeSession and unifies onto a
single session type giving access to all operations.

The XSession was renamed to Session and mysqlx.get_session() now returns
a Session object.

Test cases were changed accordingly.