BUG26029811: Improve error thrown when using an invalid parameter in bind()
The X DevAPI does not support binding anonymous parameters, but it
is possible to pass a single parameter to the bind() function in
Connector/Python. This parameter must be a JSON string or a DbDoc
object.

This patch improves the error thrown when the user is using a
parameter that is not a JSON string or DdDoc.

A test case was added for regression.