WL10754: Forbid modify() and remove() with no condition
The modify() and remove() collections operations, when called with no
condition, will apply changes to the entire collection. That is very
rarely the desired effect.
The same applies to update() and delete() in the tables.

To minimize the risk of accidental loss of data, this worklog changed
this behavior. Now for these operations the expression parameter is
mandatory and non-empty.

Tests were changed for regression.