BUG26376334: Fix prepared statements in MySQL 8.0
In MySQL 8.0, all prepared statements hang on calling `fetchall` on a cursor.
This happens because the server expects a CMD_STMT_FETCH command from the
client to send the results of the prepared query.

This patch adds new flags to the `MySQLCursorPrepared` class to check for
`SERVER_STATUS_CURSOR_EXISTS` and `SERVER_STATUS_LAST_ROW_SENT` flags in the
EOF/OK packets and then send `CMD_STMT_FETCH` command accordingly.

Tests failing with MySQL 8.0 will now pass with this patch.