WL10899: Add support for pure Python implementation of Protobuf
As of Connector/Python 2.2.3, source distributions include a mandatory C++
extension, that interfaces with a MySQL server with the X Plugin enabled
using Protobuf as data interchange format.

This worklog adds support for a pure Python implementation of Protobuf,
removing this requirement.

The API was not changed, therefore no test cases were added or changed.