WL10973: Allow JSON types as operands for IN operator
This patch updates the parsing for Json Array and Json Object in the Expression
Parser. It also adds the two new X operators: `cont_in` and `not_cont_in` to
support these operations.

The parsing of IN operator takes place as before, but we can now have an array,
dict, or a document path in the RHS of the operator to check for the single
element in the LHS.

Tests have been added for regression.