WL10053: Configuration handling interface
This patch allows the user to store/load/delete named session configurations
from a persistent storage and use these configurations to connect to
a MySQL server running the XPlugin.

In Linux, configuration files can be in /etc/mysql/sessions.json and
~/.mysql/sessions.json with the latter overwriting the former. The Windows
equivalent locations are %PROGRAMDATA%MySQLsessions.json and
%APPDATA%MySQLsessions.json

Passwords are not stored. Users must explicitly provide a password to connect.

Tests have been added for regression.