WL10975: Add Single document operations
These are commands that operate at a single document level, unlike the other
CRUD commands that operate on all documents that match a filter.

This feature adds 4 methods to `Collection`:
1. `Collection.replace_one` replaces a document identified with a document_id
   with the provided document.
2. `Collection.add_or_replace_one` does the same as replace_one, but this adds
   a new document if the document_id doesn't exist.
3. `Collection.get_one` returns a document matching a document_id.
4. `Collection.remove_one` removes a document matching a document_id.

This patch also adds the upsert(val=True) method chainable to Table.insert() and
Collection.add().

By default upsert is set to False. Setting this to True replaces the row matched
with the provided key with the values provided.

Tests have been added for regression.