WL10771: Add SHA256 authentication
This patch adds 2 new authentication modes: Plain and External.
The External authentication mechanism, at the time of writing, is not recognized
by the server and will not work.

The user can specify the authentication mode with the newly added option `auth`
which can take the following values:
1. plain or Auth.PLAIN
2. external or Auth.EXTERNAL
3. mysql41 or Auth.MYSQL41

`Auth` is a new enum added to `mysqlx.constants` module.

If the user does not specify the authentication mechanism, Plain authentication
is used if the connection mode is secure, or else Mysql41 is used.

Tests have been added for regression.