BUG26847553: Fix AttributeError when using pure Python Protobuf
Using a value of type list when setting an attribute in a Message object
throws an AttributeError.

This patch fix this issue by testing the value type and using the .extend()
method for list type values.

This patch also adds the encoding of string values types to bytes when using
Python 3.