WL11142: DevAPI: Modify/MergePatch
This Worklog provides the entry point to use the MySQL Server
Functionality to Patch JSON Documents through the addition of
a new function to the MySQL Connector Python X Developers
Application Programming Interface.

The deliverable product is a new method called "patch" that
will be added to the ModifyStatement Class (that represents a
statement for document update operations on a Collection)
that handles the functionality specified in the MYSQL X DevAPI
Specification for the Modify Collection Method. This way this
new patch function will be available to be applied to
documents resulting from a search condition applied to a
collection.

The Patch method allows the developers to add or unset
attributes in a JSON document in addition to update their
values on a single invocation. To invoke the Merge Patch
Operation, developers will require to specify a target and a
patch. The target is defined from a condition, which is
applied to a collection and all documents whom meets the
condition will become the target.

Notes:
1. Placeholder for WL#10725 for the Modify.Patch() method for
   the MySQL Connector Python X DevAPI.

v1