WL11149: Create Pylint test for mysqlx
The aim of this patch is to create a style test for the mysqlx module
using Pylint, that will enforce the coding standards and improve error
detection.

The mysqlx.errors.Warning class was removed and tuples are now being
used to store warnings from DevAPI.

A new mysqlx_style test was added.