WL11278: Collection.create_index
This Worklog provides a easier way to create the index
on a JSON document in a more JSON style way that replaces
the chained method invocation with a single method call.

The deliverable product is a change of functionality for the
current method 'create_index()', which in addition of receive
the index name it will receive a second parameter in form of
a JSON document, which will describe the index to be created.

Note: This new functionality will replace the current one
that requires the invocation of the 'field()' method, so for
the 'field()' method will be removed as part of this worklog.

Placeholder for WL#9876 for the Collection.createIndex method
for the MySQL Connector Python X DevAPI.

v1b