Introduce SSL option 'verify_identity'
In order to allow clients to verify the server name against a server
certificate, this patch adds an option 'ssl_verify_identity' to
mysql.connector that calls ssl.match_hostname(...) internally.

The option is disabled by default, so the old behaviour is unchanged.
Client code must enable the option explicitly.