BUG24948205: Results from JSON_TYPE() are returned as bytearray
Result value from running a query with JSON_TYPE() is returned
as a VAR_STRING value which is not converted and is returned as
a bytearray. With this patch the returned value will be converted
to a string type using the current charset, in case the conversion
fail the value will still returned as a bytearray value.

v2