WL11614: Enable C extension by default
This WL changes the default value of connection option use_pure from
True to False in the mysql.connector package, which enables the usage
of the C extension by default.

It also implements the mechanism for enabling the usage of pure Python
in the mysqlx package.

Tests were added for regression.