WL11448: New document _id generation support
The MySQL Connector/Python X DevAPI provides functionality to create
the id field (_id) in the documents when the user does not provide one.
However, with the introduction of WL#11390 this capability will be
available at the server side. In addition connectors shall not create
ids anymore to avoid possible collisions.
Generated ids at the server side will be available on the result of the
execution of the add statement through the method get_generated_ids(),
which returns a list.

v6