BUG27589450: Remove upsert functionality from WriteStatement class
Currently the upsert() method is part of the WriteStatement class which
is inherited by both the AddStatement and InsertStatement classes.
However, upserts are currently only supported for add statements, and
it will cause an assertion to attempt to execute an insert statement
that has enabled upserts.

This patch removes the upsert functionality from WriteStatement and adds
it to AddStatement class.