Add the restriction of _id modification in mysqlx.DbDoc.
This patch adds the restriction of _id modification in mysqlx.DbDoc.

The new method copy() returns a new copy of a mysqlx.DbDoc object
containing the doc_id provided. If doc_id is not provided, it will
be removed from the new mysqlx.DbDoc object.

A new test was created for mysqlx.DbDoc.