BUG27364914: Cursor prepared statements do not convert strings
Using the cursor with prepared statements, the returned values from
queries are converted from binaries values to their correspond python
values, like int and datetime types, but not for strings types which
bytearrays are returned.

This patch changes this behavior to return text values as string types.

A test was added for regression.

v5