BUG24948186: JSON type results are bytearray instead corresponding python type
The values returned from server that are JSON types are returned
as text (or bytearray in python 3), for example INTEGER 1 and
DOUBLE 1.5 are returned as "1" and "1.5" respectively as strings
or bytearray instead of int and float, this also applies to "BOOLEAN",
"DATETIME", "DATE", "TIME" which should be their corresponding types
of python such as datetime, date and timedelta. This happens because
on the server the JSON objects are saved in BLOB This patch fixes the
conversion for MySQL JSON types to their corresponding python types.

v3