BUG27371245: Connection fails on unsupported default's auth_plugin
The connection to a server configured with an unsupported
authorization plugin fails in all cases, even if the user does not
require such plugin to authenticate himself. This is because the
connector raises an error if the server's default auth_plugin is
not recognized or unsupported.

This patch changes this behavior to fall back to a default
auth_plugin supported by the connector in the case the server's
default auth_plugin is not recognized by the connector.

Note: the "mysql_native_password" has been set the default
auth_plugin for the connector. But in case the user requires a
different one, the server will send an AuthSwitchRequest to use a
different plugin.