BUG27650437: Fix differences between pure Python and CEXT for get_row()/get_rows()
While using and comparing the returned results from a call to
get_row and get_rows from the pure python implementation and
the c-extension some discrepancies on the format are notorious
between them, in addition the included status flag and warning
are missing from the results from the c-extension.

This patch corrects the differences on the results from the
c-extension.

v3