BUG27991948: unread_result is not unset after invoke get_rows on c-ext
The invocation of cmd_query after retrieve the results of a previews
query with get_rows raise an InternalError: "Unread result found"
this is because the unread_result property is not automatically set
to false after the results are read.

This patch improves the usability of the c-extension by unseting
the unread_result property to mimic the behavior of the pure
python implementation that set to false this variable when the
results had been read, with this change the user will not
require to invoke the free_result method in order of reuse
the same connection to run a new query.