BUG26834200: Deprecate Row.get_string() method
The method Row.get_string() is used to retrieve a value from a row
by the column name.
This patch deprecates Row.get_string() and implements the funcionality
using Row.__getitem__(), which allows access to the value from a row
by the column name or index.

Tests were added for regression.