Fix unittest failures when stopping the server
The stopping server mechanism fails due to the wrong PID used to
stop the server, the PID used is the one from the process that
start the server, instead of the mysqld process itself.

This patch changes the wrong behavior to always get the PID from
the PID file to stop the server. In addition this patch adds new
logging entries to facilitate the debug of unittest that requires
to stop the server.