BUG28278352: _mysqlxpb python objects are not deleted
While using mysqlx with the c extension, a repeated use of
the collection.add() and collection.find() creates too much
objects causing to use more memory that requires.

This patch decreases the Python objects created on _mysqlxpb
using the Py_CLEAR macros to remove the references to the objects
once they are unused.