BUG26484601: unable to connect to a server using other than TLSv1
The connection to a server that is configured to use a different
version of TLSv1 (as TLSv1.1 and TLSv1.2) is not possible, and
ends with "system error: 1 [SSL: WRONG_VERSION_NUMBER] wrong
version number (_ssl.c:590)". This happens because the connector
is wrongly restricting the connection to use only TLSv1.

This patch changes the behavior to allow the server to specify
which version should be used in the secured connection. However
the the connector can still restrict which version it wants to
use but this should be used for testing purposes since is more
secure to leave the server to choose the version, who naturally
chooses the latest available between the versions allowed in
tsl_version configuration option.