BUG28443941: differences on pure and c-ext in cmd_change_user()
The invocation of the method cmd_change_user() using the c extension
does not return any status information while the pure python
implementation it does.

This patch adds the missing status results for the cmd_change_user(),
in addition this patchs renames the server_status atribute name to
status_flag from the status information returned from fetch_eof_status()
to be consistent with the name.