WL11897: Implement connection pooling for xprotocol
This worklog will add the functionality of a connection pooling
to the Python MySQLx dev API. The deliverable for this worklog
is a new method get_client() to the mysqlx module that will
receive a connection information, and the settings for the
connections. This method will return a new object of class
Client from where new Session instances can be obteined. A pool
mechanism for connections for the session can be customized by
get_client().