Fix unittest bad authentication hash
In some occasions the server 5.6.40+ sends an authentication data
for the Native password that once is scrambled along the user and
password will produce an invalid authentication message.

This patch verify the authentication message in search of additional
parameter separators \x00 which corrupts the message and cause the
authentication to fail. In such cases the connection attempt will be
restarted, by requesting new authentication data to produce a new
authentication message.