BUG27434751: Add a TLS/SSL option to verify server name
To prevent man-in-the-middle attacks, MySQL clients should connect using TLS
and verify the server name against the server certificate's common name (CN)
and subject alternative names (SANs).

This patch adds a new connection option `ssl_verify_identity` to perform this
verification in the pure Python implementation, and also changes the behavior
of the C extension implementation, which previously was performing this
verification by default.

A test was added for regression.