WL12488: Support for session-connect-attributes
The purpose of this Worklog is to support session-connect-attributes,
there are two different kinds of session attributes:

Client-defined attributes, created by the client connected to the
server that will be defined by this Python X DevAPI Implementation
which should be immutable during the life-time of the session.

User-defined attributes that are key-value mappings provided by
the user/application via the public API at the time the a new
Session is object created.