BUG27897881: empty LONG BLOB throws an IndexError
Quering an empty long blob raises an IndexError.
This happens because the JSON values are stored in BLOBS
which must be tested against other data types to get the
python equivalent type, but one of the conversion method
raises the IndexError while trying to convert the data.
This patch fixes the wrong behavior avoiding leaving the
exception unhanled and by fixing the data types conversion
methods.

Reviewer: Nuno Mariz
http://rb.no.oracle.com/rb/r/21074/