WL12489: DevAPI: Support new session reset functionality
The Pooling implementation supported on MySQL 8.0.13 is limited
by the server side; who closes the connection after the client
asks for reset the session. MySQL 8.0.16 provides full
functionality; connections will not be closed by the server after
they are reset by the client. This worklog ensures the Python
Dev API supports this new server functionality on the client'sside.