BUG29855733: Fix error during connection using charset and collation combination
An error is raised when using the combination of utf8mb4 as charset and
utf8mb4_0900_bin as collation. The character id of this combination is 309,
which is greater than 1 byte specified in the handshake protocol.

This patch uses 1 extra byte from the reserved part of the payload, to
accommodate the necessary 2 bytes of this combination.

Tests were added/changed for regression.