Fix new ssl libraries on msi and failing unitest on Py2.7
The ssl libraries "ssleay32.dll" and "libeay32.dll" had change for
the "libssl-1_1.dll" and "libcrypto-1_1.dll" on 32bit msi packages
and "libssl-1_1-x64.dll" and "libcrypto-1_1-x64.dll" for the 64bit
msi packages. This patch made the adecuate changes to include these
new libraries.

In addition this patch avoids a timeout issue on the tests:
  bugs.Bug865859.test_reassign_connection (using MySQLConnection)
  bugs.BugOra18144971.test_prepared_statement (using CMySQLConnection)
On el7 with Pyv2.7 generated by the opening of a subprocess to load
the distribution information from lsb_release.