fix CMySQLCursor.fetchmany bug
Issue: When calling fetchmany iteratively and reaching end of rows in cursor, fetchmany does not updates self. _nextrow if there is no unread_result (only happens when the last batch is not a full batch).
Fix: add an additional if/else where the self._nextrow is reset to (None, None) when unread_result is False.