WL13334: Connector support for failover and load-balancing
With this worklog Connector Python will support multiple remote instances of
the Router, by allowing the developer to specify more than one address to
connect to.

The Connector will randomly pick one of the given servers to attempt to connect
to it. In case the connection fails, it will try another. finally if all
addresses fails to connect to, it will raise a connection error.