WL13372: DNS SRV support
A SRV record is a specification of data in the DNS, that typically
defines a symbolic name and the transport protocol used as part of
the domain name.
It defines the priority, weight, port, and target for the service
in the record content.

This worklog implements the DNS SRV records support, which allows
resolving SRV records available in a DNS server.

The implementation of routers in the X DevAPI and failover in
classic protocol were changed to support SRV records.

The `dnspython` package was added as a required dependency.