WL13847: Add support for context managers
Context managers in Python, allows the allocation and releasing of
resources precisely when it's needed.
The most widely used example of context managers is the with statement.

This worklog implements this functionality in the classic protocol for
Connection objects and in the X DevAPI protocol for Session and Client
objects.