BUG31098686: unittests for cmd_change_user does not cover CMySQLConnection defaults.
The existing unittest for cmd_change_user for the cext does not uses the defaults,
in addition it does not run with all the server version currently supported. This patch
updates the unittest for cmd_change_user to run over 5.6, 5.7 and 8.0 with a greater set
of different authentification users.

Note: Documentation needs to be updated, In order to successfuly run cmd_change_user the
option 'charset' must be set to 'utf8' when the connector is used with MySQL 5.6 and 5.7
due to the default 'utf8mb4' is only supported on MySQL 8.0.