BUG27535063: Fix wrong error message when specifying the collation for a non TEXT field
When creating an index on a collection, if a collation is specified but the
field is not of the type TEXT, then an error message is generated but
with a wrong field type.

This patch fixes the issue by setting the correct field type. Previously
it always raised with "GEOJSON" for all field types.

Tests were added for regression.