WL14213: Support GSSAPI - Kerberos auth
The purpose of this Worklog is to add support for the SASL authentication
protocol using the GSSAPI (Kerberos) authentication method for the pure
python implementation.

This changes require the GSSAPI pypi module https://pypi.org/project/gssapi/
which provides both low-level and high level wrappers around the GSSAPI C
libraries. The SASL mechanism is handled by C/py in connection module.

The GSSAPI pypi module requires MIT kerberos installed in the system in
order to work and be able to request tickets to authentificate the user with
the MySQL server when the user is IDENTIFIED WITH authentication_ldap_sasl
and the authentication_ldap_sasl plugin is configured to use the "GSSAPI"
mechanism.